% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutbar.R
\name{scoutbar}
\alias{scoutbar}
\alias{update_scoutbar}
\title{Creates a Scoutbar React widget}
\usage{
scoutbar(
  inputId,
  theme = c("light", "dark", "auto"),
  placeholder = list("Hello", "Type some text"),
  actions = list(),
  ...
)

update_scoutbar(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  configuration = NULL
)
}
\arguments{
\item{inputId}{Widget input id.}

\item{theme}{Scoutbar theme.}

\item{placeholder}{Scoutbar placeholder text. A string or
a list of strings.}

\item{actions}{Scoutbar actions. Expect \link{scout_action} or \link{scout_section} and
\link{scout_page}. \link{scout_action} can be nested inside \link{scout_section}
and \link{scout_page}.}

\item{...}{Any other configuration parameter. See
\url{https://www.scoutbar.co/docs/features}.}

\item{session}{Shiny session object.}

\item{configuration}{Scoutbar configuration. Expect a list of properties.
See possible values here at \url{https://www.scoutbar.co/docs/features}.}
}
\description{
Scoutbar react widget for Shiny.

Usee this function from the server side of
your Shiny app.
}
