% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutbar.R
\name{scout_action}
\alias{scout_action}
\title{Creates a scout action}
\usage{
scout_action(id, label, description, closeOnClick = TRUE, ...)
}
\arguments{
\item{id}{Unique id.}

\item{label}{Action label.}

\item{description}{Action description.}

\item{closeOnClick}{Whether to close the scoutbar whenever this action is
clicked. Default to TRUE.}

\item{...}{Other options. See \url{https://www.scoutbar.co/docs/actions}.}
}
\description{
Can sort \link{scout_action} on the same view.
}
