% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutbar.R
\name{scout_page}
\alias{scout_page}
\title{Creates a scout page}
\usage{
scout_page(label, ..., .list = NULL)
}
\arguments{
\item{label}{Page label.}

\item{...}{Expect \link{scout_action}.}

\item{.list}{To pass a list of \link{scout_action}.}
}
\description{
Can embed \link{scout_action} on a separate
view of the Scoutbar.
}
