% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutbar.R
\name{scout_section}
\alias{scout_section}
\title{Creates a scout section}
\usage{
scout_section(label, ..., .list = NULL)
}
\arguments{
\item{label}{Section label.}

\item{...}{Expect \link{scout_action}.}

\item{.list}{To programmatically pass a list of \link{scout_action},
with \code{lapply} for instance.}
}
\value{
A list containing:
\itemize{
\item children: a sublist where are passed the \link{scout_action}.
\item label: The section label.
\item class: a character vector to identify the section on the JavaScript side.
You are not expected to modify it as it will break the JavaScript binding.
}
}
\description{
Sort \link{scout_action} on the same view.
}
\details{
Whenever many \link{scout_action} share a similar topic,
you may use this function to sort them in the UI and offer
a better user experience. You can combine it with \link{scout_page}.
}
