% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoutbar.R
\name{scout_section}
\alias{scout_section}
\title{Creates a scout section}
\usage{
scout_section(..., .list = NULL, label)
}
\arguments{
\item{...}{Expect \link{scout_action}.}

\item{.list}{To pass a list of \link{scout_action}.}

\item{label}{Section label.}
}
\description{
Can sort \link{scout_action} on the same view.
}
